.travis/requirements.txt
Jinja2
cairosvg
Pillow
requests